requirements/base.txt
astroid==1.6.3
autopep8==1.3.5
bcrypt==3.1.4
blinker==1.4
cffi==1.11.5
click==6.7
flake8==3.5.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Moment==0.6.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
Pillow==5.1.0
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
pylint==1.8.4
python-dotenv==0.8.2
six==1.11.0
SQLAlchemy==1.2.7
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.1
